Linting and formatting to PR review standards
