+++ src/cashier_frontend_new/src/modules/creationLink/types/index.ts
@@ -1,6 +1,14 @@
 import type { GatingStore } from "$modules/gating/state/gatingStore.svelte";
 import type { TransactionLockInput } from "$modules/gating/types/transactionLockDisplay";
 
+export type {
+  DraftLink,
+  DraftLinkCreateParams,
+  DraftLinkStorageRecord,
+  DraftLinkUpdateData,
+  DraftLinkUpdateParams,
+} from "$modules/creationLink/types/draftLink";
+
 export type TransactionLocksDrawerProps = {
   open?: boolean;
   locks?: TransactionLockInput[];
